Gonna need some help from you on this one.
I figured out that audio is what is slowing us down so I took the audio file
and put it on it's own empty game object. I used the dont destroy on load
feature on this object to make sure it persist which clears all our load time issues.
Only problem is that if a player returns to the main menu it creates another
instance of the audio so it plays two tracks at once.

I tried doing some singleton method stuff after googling but couldn't get it to work
Maybe you can take a look Hank?